// autogpt_platform/frontend/src/app/api/__generated__/models/userTransaction.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { CreditTransactionType } from "./creditTransactionType";
import type { UserTransactionDescription } from "./userTransactionDescription";
import type { UserTransactionUsageGraphId } from "./userTransactionUsageGraphId";
import type { UserTransactionUsageExecutionId } from "./userTransactionUsageExecutionId";
import type { UserTransactionUserEmail } from "./userTransactionUserEmail";
import type { UserTransactionReason } from "./userTransactionReason";
import type { UserTransactionAdminEmail } from "./userTransactionAdminEmail";
import type { UserTransactionExtraData } from "./userTransactionExtraData";

export interface UserTransaction {
  transaction_key?: string;
  transaction_time?: Date;
  transaction_type?: CreditTransactionType;
  amount?: number;
  running_balance?: number;
  current_balance?: number;
  description?: UserTransactionDescription;
  usage_graph_id?: UserTransactionUsageGraphId;
  usage_execution_id?: UserTransactionUsageExecutionId;
  usage_node_count?: number;
  usage_start_time?: Date;
  user_id: string;
  user_email?: UserTransactionUserEmail;
  reason?: UserTransactionReason;
  admin_email?: UserTransactionAdminEmail;
  extra_data?: UserTransactionExtraData;
}